+++ types/types.go
@@ -1,0 +1,5 @@
+// Package types contains types that are common across libnetwork project
+package types
+
+// UUID represents a globally unique ID of various resources like network and endpoint
+type UUID string
